refactor: replace SqliteKnowledgeRelationalStore with SqliteRelationalGateway in test harness and related components
- Updated references in test harness to utilize the new SqliteRelationalGateway for loading repo IDs and production artefacts.
- Introduced new methods in the RelationalGateway trait for loading artefacts and repo IDs.
- Cleaned up unused methods and types related to the previous SqliteKnowledgeRelationalStore.
- Enhanced test cases to align with the new relational gateway structure.

Made-with: Cursor
