+++ bitloops/src/capability_packs/test_harness/storage/dispatch.rs
@@ -10,11 +10,10 @@ use crate::config::{RelationalProvider, resolve_store_backend_config_for_repo};
 use crate::models::{
     CoverageCaptureRecord, CoverageDiagnosticRecord, CoverageHitRecord, CoveragePairStats,
     CoverageSummaryRecord, CoveringTestRecord, LatestTestRunRecord, ListedArtefactRecord,
-    ProductionArtefact, ProductionIngestionBatch, QueriedArtefactRecord,
-    ResolvedTestScenarioRecord, StageBranchCoverageRecord, StageCoverageMetadataRecord,
-    StageCoveringTestRecord, StageLineCoverageRecord, TestDiscoveryDiagnosticRecord,
-    TestDiscoveryRunRecord, TestHarnessCommitCounts, TestLinkRecord, TestRunRecord,
-    TestScenarioRecord, TestSuiteRecord,
+    ProductionIngestionBatch, QueriedArtefactRecord, ResolvedTestScenarioRecord,
+    StageBranchCoverageRecord, StageCoverageMetadataRecord, StageCoveringTestRecord,
+    StageLineCoverageRecord, TestDiscoveryDiagnosticRecord, TestDiscoveryRunRecord,
+    TestHarnessCommitCounts, TestLinkRecord, TestRunRecord, TestScenarioRecord, TestSuiteRecord,
 };
 use crate::storage::init::init_test_domain_database;
 
@@ -75,39 +74,10 @@ pub fn open_repository_for_repo(repo_root: &Path) -> Result<BitloopsTestHarnessR
 }
 
 impl TestHarnessRepository for BitloopsTestHarnessRepository {
-    fn load_repo_id_for_commit(&self, commit_sha: &str) -> Result<String> {
-        match self {
-            Self::Sqlite(repository) => repository.load_repo_id_for_commit(commit_sha),
-            Self::Postgres(repository) => repository.load_repo_id_for_commit(commit_sha),
-        }
-    }
-
-    fn load_production_artefacts(&self, commit_sha: &str) -> Result<Vec<ProductionArtefact>> {
-        match self {
-            Self::Sqlite(repository) => repository.load_production_artefacts(commit_sha),
-            Self::Postgres(repository) => repository.load_production_artefacts(commit_sha),
-        }
-    }
-
     fn load_test_scenarios(&self, commit_sha: &str) -> Result<Vec<ResolvedTestScenarioRecord>> {
         match self {
             Self::Sqlite(repository) => repository.load_test_scenarios(commit_sha),
             Self::Postgres(repository) => repository.load_test_scenarios(commit_sha),
-        }
-    }
-
-    fn load_artefacts_for_file_lines(
-        &self,
-        commit_sha: &str,
-        file_path: &str,
-    ) -> Result<Vec<(String, i64, i64)>> {
-        match self {
-            Self::Sqlite(repository) => {
-                repository.load_artefacts_for_file_lines(commit_sha, file_path)
-            }
-            Self::Postgres(repository) => {
-                repository.load_artefacts_for_file_lines(commit_sha, file_path)
-            }
         }
     }
 
@@ -191,18 +161,6 @@ impl TestHarnessRepository for BitloopsTestHarnessRepository {
 }
 
 impl TestHarnessCoverageGateway for BitloopsTestHarnessRepository {
-    fn load_repo_id_for_commit(&self, commit_sha: &str) -> Result<String> {
-        TestHarnessRepository::load_repo_id_for_commit(self, commit_sha)
-    }
-
-    fn load_artefacts_for_file_lines(
-        &self,
-        commit_sha: &str,
-        file_path: &str,
-    ) -> Result<Vec<(String, i64, i64)>> {
-        TestHarnessRepository::load_artefacts_for_file_lines(self, commit_sha, file_path)
-    }
-
     fn insert_coverage_capture(&mut self, capture: &CoverageCaptureRecord) -> Result<()> {
         TestHarnessRepository::insert_coverage_capture(self, capture)
     }
