+++ bitloops/src/host/capability_host/gateways.rs
@@ -1,6 +1,7 @@
 pub mod blob_payloads;
 pub mod documents;
 pub mod relational;
+pub mod sqlite_relational;
 
 use anyhow::Result;
 use serde_json::Value;
@@ -11,6 +12,7 @@ pub use crate::adapters::connectors::{
 pub use blob_payloads::BlobPayloadGateway;
 pub use documents::DocumentStoreGateway;
 pub use relational::RelationalGateway;
+pub use sqlite_relational::SqliteRelationalGateway;
 
 pub trait CanonicalGraphGateway: Send + Sync {}
