+++ bitloops/src/host/capability_host/gateways/relational.rs
@@ -3,6 +3,7 @@ use anyhow::Result;
 use crate::capability_packs::knowledge::storage::{
     KnowledgeItemRow, KnowledgeRelationAssertionRow, KnowledgeSourceRow,
 };
+use crate::models::ProductionArtefact;
 
 /// Host relational port for capability packs. The default implementation backs the Knowledge pack
 /// schema on the DevQL relational SQLite file; other packs use pack-scoped tables and migrations.
@@ -21,4 +22,12 @@ pub trait RelationalGateway: Send + Sync {
     fn list_items_for_repo(&self, repo_id: &str, limit: usize) -> Result<Vec<KnowledgeItemRow>>;
     fn resolve_checkpoint_id(&self, repo_id: &str, checkpoint_ref: &str) -> Result<String>;
     fn artefact_exists(&self, repo_id: &str, artefact_id: &str) -> Result<bool>;
+
+    fn load_repo_id_for_commit(&self, commit_sha: &str) -> Result<String>;
+    fn load_production_artefacts(&self, commit_sha: &str) -> Result<Vec<ProductionArtefact>>;
+    fn load_artefacts_for_file_lines(
+        &self,
+        commit_sha: &str,
+        file_path: &str,
+    ) -> Result<Vec<(String, i64, i64)>>;
 }
